+++ grafanalib/tests/test_core.py
@@ -1088,6 +1088,27 @@ def test_ae3e_plotly():
     assert data["options"]["layout"] == layout
 
 
+def test_barchart():
+    data_source = "dummy data source"
+    targets = ["dummy_prom_query"]
+    title = "dummy title"
+    panel = G.BarChart(data_source, targets, title)
+    data = panel.to_json_data()
+    assert data["targets"] == targets
+    assert data["datasource"] == data_source
+    assert data["title"] == title
+    assert data["options"] is not None
+    assert data["fieldConfig"] is not None
+    assert data["options"]["orientation"] == 'auto'
+    assert data["fieldConfig"]["defaults"]["color"]["mode"] == 'palette-classic'
+
+    panel = G.BarChart(data_source, targets, title, orientation='horizontal', axisCenteredZero=True, showLegend=False)
+    data = panel.to_json_data()
+    assert data["options"]["orientation"] == 'horizontal'
+    assert data["fieldConfig"]["defaults"]["custom"]["axisCenteredZero"]
+    assert not data["options"]["legend"]["showLegend"]
+
+
 def test_target_invalid():
     with pytest.raises(ValueError, match=r"target should have non-empty 'refId' attribute"):
         return G.AlertCondition(