+++ grafanalib/tests/test_core.py
@@ -1135,3 +1135,31 @@ def test_sql_target():
     )
     assert t.to_json_data()["targets"][0].rawQuery is True
     assert t.to_json_data()["targets"][0].rawSql == "SELECT * FROM example"
+
+
+def test_sql_target_with_source_files():
+    t = G.Table(
+        dataSource="some data source",
+        targets=[
+            G.SqlTarget(srcFilePath="grafanalib/tests/examples/sqltarget_example_files/example.sql"),
+        ],
+        title="table title",
+    )
+    assert t.to_json_data()["targets"][0].rawQuery is True
+    assert t.to_json_data()["targets"][0].rawSql == "SELECT example, count(id)\nFROM test\nGROUP BY example;\n"
+    print(t.to_json_data()["targets"][0])
+
+    t = G.Table(
+        dataSource="some data source",
+        targets=[
+            G.SqlTarget(srcFilePath="grafanalib/tests/examples/sqltarget_example_files/example_with_params.sql", sqlParams={
+                "example": "example",
+                "starting_date": "1970-01-01",
+                "ending_date": "1971-01-01",
+            },),
+        ],
+        title="table title",
+    )
+    assert t.to_json_data()["targets"][0].rawQuery is True
+    assert t.to_json_data()["targets"][0].rawSql == "SELECT example\nFROM test\nWHERE example='example' AND example_date BETWEEN '1970-01-01' AND '1971-01-01';\n"
+    print(t.to_json_data()["targets"][0])