+++ CHANGELOG.rst
@@ -12,9 +12,10 @@ Changelog
 * Added Bar_Chart_ panel support
 * Extended SqlTarget to support parsing queries from files
 * Fix AlertCondition backwards compatibility (``useNewAlerts`` default to ``False``)
+* Added RateMetricAgg_ for ElasticSearch
 
 .. _`Bar_Chart`: https://grafana.com/docs/grafana/latest/panels-visualizations/visualizations/bar-chart/
-
+.. _`RateMetricAgg`: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-metrics-rate-aggregation.html
 
 0.7.0 (2022-10-02)
 ==================